Refactor SkillsSettings component and styles; move to new directory structure
