+++ client/src/pages/openPositions/OpenPositions.jsx
@@ -8,7 +8,7 @@ import { UseUser } from "../../context/UserContext";
 import "./OpenPositions.css";
 import { findFilterOptions, filterJobs } from "../../util/filterJobs";
 import getSkillsInDescription from "../../util/getSkillsInDescription";
-import SkillsSettings from "../../components/SkillsSettings";
+import SkillsSettings from "../../components/SkillsSettings/SkillsSettings";
 import { UseJobs } from "../../context/JobsContext.jsx";
 import createSortComparator from "../../util/createSortComparator";
